requirements.txt
pymongo>=4.5.0
python-dotenv>=1.0.0